+++ acceptance-tests/apps/postgresqlapp/internal/app/app.go
@@ -36,36 +36,6 @@ func App(uri string) *mux.Router {
 	return r
 }
 
-func handleAlterTable(db *sql.DB) func(http.ResponseWriter, *http.Request) {
-	return func(w http.ResponseWriter, r *http.Request) {
-		log.Println("Handling public.test table alteration.")
-
-		_, err := db.Exec(fmt.Sprintf(`ALTER TABLE public.test alter column %s type varchar(256);`, valueColumn))
-		if err != nil {
-			fail(w, http.StatusBadRequest, "Error altering the table %v", err)
-			return
-		}
-
-		w.WriteHeader(http.StatusOK)
-		log.Printf("table public.test modified")
-	}
-}
-
-func handleDeleteTestTable(db *sql.DB) func(http.ResponseWriter, *http.Request) {
-	return func(w http.ResponseWriter, r *http.Request) {
-		log.Println("Handling public.test table deletion.")
-
-		_, err := db.Exec(`DROP TABLE public.test`)
-		if err != nil {
-			fail(w, http.StatusBadRequest, "Error dropping table public.test %v", err)
-			return
-		}
-
-		w.WriteHeader(http.StatusNoContent)
-		log.Printf("table public.test dropped")
-	}
-}
-
 func aliveness(w http.ResponseWriter, _ *http.Request) {
 	log.Printf("Handled aliveness test.")
 	w.WriteHeader(http.StatusNoContent)
